Source/requirements.txt
aiohttp
websockets
protobuf
PyYAML
aiosqlite